% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpine_skiing}
\alias{alpine_skiing}
\title{Example Alpine Skiing}
\format{A data frame with 4707 rows and 8 variables:
\describe{
  \item{pointID}{ID for each track point}
  \item{ele}{elevation above sealevel in meters}
  \item{time}{time (in POSIXct-format)}
  \item{runtime}{time since start (in seconds)}
  \item{distance}{distance since the last trackpoint}
  \item{cumulative.distance}{cumulative distance since the beginnning}
  \item{x}{x-coordinate (longitude (East-West-Dimension))}
  \item{y}{y-coordinate (longitude (North-South-Dimension))}
}}
\usage{
alpine_skiing
}
\description{
An alpine skiing afternoon in St. Moritz, Switzerland
}
\keyword{datasets}
