% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind}
\alias{bind}
\title{Bind parameters to a function}
\usage{
bind(func, ...)
}
\arguments{
\item{func}{Function prototype object \strong{required}.}

\item{...}{Unnamed or named parameters to bind \strong{required}.}
}
\description{
\code{bind} returns a closure with given parameters bound to it.
}
\examples{
boo <- function(a, b) a + b  # func with no defaults
runSeries(list(function() 0L, 
               bind(boo, 1L, 2L)))  # bind on the fly

}
\seealso{
\code{\link{runSeries}} \code{\link{runWaterfall}} 
\code{\link{runRace}} \code{\link{runParallel}}
}

