% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc}
\alias{calc}
\alias{calc_pace}
\alias{calc_splits}
\alias{calc_time}
\alias{calc_dist}
\title{Calculate Pace, Time, Distance, and Splits}
\usage{
calc_pace(distance, time)

calc_splits(distance, time)

calc_time(distance, pace)

calc_dist(time, pace)
}
\arguments{
\item{distance}{numeric distance traveled (Units irrelavant)}

\item{time}{character time vector "hh:mm:ss" or "mms:ss"}

\item{pace}{character pace vector "hh:mm:ss" or "mms:ss"}
}
\value{
`calc_pace()` and `calc_time()` return character vectors.
`calc_distance()` returns a numeric vector. `calc_splits()` return a data.frame.
}
\description{
Inspired by https://davidrroberts.wordpress.com/2016/04/04/running-pace-and-time-calculator-function-for-r/
}
\examples{

dist <- 26.2

pace <- "08:01"

time <- "03:30:00"

calc_time(dist, pace)
calc_pace(dist, time)
calc_dist(time, pace)
calc_splits(dist, time)
}
