% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{convert_dist}
\alias{mi_to_km}
\alias{km_to_mi}
\alias{convert_pace}
\title{Converts distance and pace units to and from different metrics}
\usage{
convert_dist(x, from, to)

mi_to_km(x)

km_to_mi(x)

convert_pace(x, from, to)
}
\arguments{
\item{x}{distance or pace to convert}

\item{from}{metric abbreviation to convert from}

\item{to}{metric abbreviation to convert to}
}
\value{
numeric or character vector
}
\description{
Converts distance and pace units to and from different metrics
}
\details{
`mi_to_km()` and `km_to_mi()` are wrappers around `convert_dist()`,
since those are the most commonly used running units.
}
\examples{

five_k <- 5

pace <- "08:30"

convert_dist(five_k, from = 'km', to = 'mi')

km_to_mi(five_k)

convert_pace(pace, from = 'mi', to = 'km')

}
