% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_sha1}
\alias{file_sha1}
\title{SHA1 digest of a file's contents}
\usage{
file_sha1(file)
}
\arguments{
\item{file}{path to the file to hash.}
}
\value{
The SHA1 digest of the file contents, as a string.
}
\description{
SHA1 digest of a file's contents
}
