% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fn_running}
\alias{fn_running}
\title{Example 5k run.}
\format{A data frame with 500 rows and 7 variables:
\describe{
  \item{pointID}{ID for each track point}
  \item{ele}{elevation above sealevel in meters}
  \item{time}{time (in POSIXct-format)}
  \item{distance}{distance since the last trackpoint}
  \item{cumulative.distance}{cumulative distance since the beginnning}
  \item{x}{x-coordinate (longitude (East-West-Dimension))}
  \item{y}{y-coordinate (longitude (North-South-Dimension))}
}}
\usage{
fn_running
}
\description{
A 5k run I did in my current university town Friedrichshafen, Germany.
}
\keyword{datasets}
