% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_logfolder}
\alias{import_logfolder}
\title{Import log-file folder}
\usage{
import_logfolder(folderpath = "data", type = "gpx", track_progress = TRUE)
}
\arguments{
\item{folderpath}{path of the folder containing the files.}

\item{type}{Type of log-files: currently, only gpx is supported.}

\item{track_progress}{T/F if data loading progress tracking is wished for}
}
\value{
A list containing a data frame for each logfile. Each such data frame contains: 
\describe{
  \item{pointID}{ID for each track point}
  \item{ele}{elevation above sealevel in meters}
  \item{time}{time (in POSIXct-format)}
  \item{distance}{distance since the last trackpoint}
  \item{cumulative.distance}{cumulative distance since the beginnning}
  \item{x}{x-coordinate (longitude (East-West-Dimension))}
  \item{y}{y-coordinate (longitude (North-South-Dimension))}
}
}
\description{
Import a folder of log files to a list.
}
\section{Remark}{

So far, only the gpx-format from Sportstracker is implemented.
}

\examples{
\dontrun{
runlist <- import_logfolder(folderpath = 'data',type = 'gpx')
head(test[[1]])
}
}
