% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_run}
\alias{import_run}
\title{Import Run}
\usage{
import_run(file = "data/2017-11-06.gpx", type = "gpx",
  track_progress = TRUE)
}
\arguments{
\item{file}{path and file name.}

\item{type}{Type of log-file: currently, only gpx is supported.}

\item{track_progress}{T/F if data loading progress tracking is wished for}
}
\value{
A data frame containing basic information about your run. 
\describe{
  \item{pointID}{ID for each track point}
  \item{ele}{elevation above sealevel in meters}
  \item{time}{time (in POSIXct-format)}
  \item{runtime}{time since start (in seconds)}
  \item{distance}{distance since the last trackpoint}
  \item{cumulative.distance}{cumulative distance since the beginnning}
  \item{x}{x-coordinate (longitude (East-West-Dimension))}
  \item{y}{y-coordinate (longitude (North-South-Dimension))}
}
}
\description{
Import Run (Standard: GPX) to Data Frame
}
\section{Remark}{

So far, only the gpx-format from Sportstracker is implemented.
}

\examples{
\dontrun{
run <- import_run(file = 'data/2017-11-06.gpx',type = 'gpx')
head(run)
}
}
