% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_s3_dispatch_method}
\alias{is_s3_dispatch_method}
\title{Does a function dispatch on S3?}
\usage{
is_s3_dispatch_method(fun)
}
\arguments{
\item{fun}{the function to inspect.}
}
\value{
\code{TRUE} if \code{fun} calls \code{UseMethod()} or \code{NextMethod()}, i.e. if it is
an S3 generic or an S3 method that delegates further.
}
\description{
Does a function dispatch on S3?
}
