% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSession.R
\name{plotSession}
\alias{plotSession}
\title{Plot running session dashboard.}
\usage{
plotSession(session)
}
\arguments{
\item{session}{Running session}
}
\description{
Plot running session dashboard.
}
\examples{
s <- runSession(sample(seq(10, 13, by = .2), 10, replace = T), rnorm(10, mean = 150))
plotSession(s)
}

