% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_file}
\alias{read_file}
\title{Read a whole file into a single string}
\usage{
read_file(filename)
}
\arguments{
\item{filename}{path to the file to read.}
}
\value{
The file contents as a length-one character vector, newlines
included.
}
\description{
Read a whole file into a single string
}
