% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPX.R
\name{readGPX}
\alias{readGPX}
\title{Function to read in .gpx file and calculate time and distance}
\usage{
readGPX(filepath)
}
\arguments{
\item{filepath}{Full path to .gpx file}
}
\value{
sf object with added columns relating to distance and time - see description
}
\description{
Function to read in .gpx file and calculate time and distance
}
