% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevPlot.R
\name{run2elev}
\alias{run2elev}
\title{run2elev}
\usage{
run2elev(data = alpine_skiing, color = "black", smooth = TRUE,
  smoothfactor = 1, smoothcolor = "red")
}
\arguments{
\item{data}{A run data frame returned by \code{import_run}.}

\item{color}{The color to draw the movement line in.}

\item{smooth}{Logical T/F whether to add a smooth function.}

\item{smoothfactor}{Loess smoothing factor}

\item{smoothcolor}{Color to draw the smoothline in.}
}
\value{
A ggplot object.
}
\description{
Visualize a run/hike temporal elevation.
}
\section{Integration into ggplot2}{

You can simply add to a plot created by \code{run2elev} by using the usual \code{ggplot2}-syntax.
}

\examples{
data("alpine_skiing")

run2elev(data = alpine_skiing)

run2elev(data = alpine_skiing,
         color = "black",
         smooth = TRUE,
         smoothfactor = 0.3,
         smoothcolor = "red") 
         
run2elev(data = alpine_skiing) + 
geom_hline(yintercept = mean(alpine_skiing$ele,na.rm = TRUE),
           color = "green")    
}
