% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlot.R
\name{run2map}
\alias{run2map}
\title{Run2Map}
\usage{
run2map(data = cc_skiing, ...)
}
\arguments{
\item{data}{A run data frame returned by \code{import_run}.}

\item{...}{Further Options passed into Leaflet's \code{addPolylines} function.}
}
\value{
A leaflet object.
}
\description{
Visualize Run2Map
}
\section{Note}{

You may customize this plot by either specifying more options within the function or adding other leaflet layers afterwards.
}

\examples{
data(data = cc_skiing)

run2map(data = cc_skiing)

run2map(data = cc_skiing,
        color = "red",
        smoothFactor = 1)

# Add circle at starting point
run2map(data = cc_skiing) + 
addCircles(lng = cc_skiing$x[1],
           lat = cc_skiing$y[1],
           popup = "START")
}
