% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProfile.R
\name{runProfile}
\alias{runProfile}
\title{Create a profile for runner.}
\usage{
runProfile(heart.rate.min = NA, heart.rate.max = NA,
  gender = .gender.masculin, age = NA, weight = NA)
}
\arguments{
\item{heart.rate.min}{Minimum heart rate}

\item{heart.rate.max}{Maximum heart rate}

\item{gender}{Gender}

\item{age}{Age}

\item{weight}{Weight in kilograms}
}
\value{
Runner's profile.
}
\description{
Create a profile for runner.
}
\examples{
runProfile(50, 175)
}

