% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runr-package.R
\docType{package}
\name{runr-package}
\alias{runr}
\alias{runr-package}
\title{runr: Run R Experiments}
\description{
Helper functions and driver scripts for running large-scale, corpus-wide R
experiments: extracting runnable code from packages, running it under a
controlled R process, and reading back the results of GNU parallel jobs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PRL-PRG/runr}
  \item Report bugs at \url{https://github.com/PRL-PRG/runr/issues}
}

}
\author{
\strong{Maintainer}: Filip Krikava \email{krikava@gmail.com} (\href{https://orcid.org/0000-0002-0478-6202}{ORCID})

Authors:
\itemize{
  \item Filip Krikava \email{krikava@gmail.com} (\href{https://orcid.org/0000-0002-0478-6202}{ORCID})
}

}
\keyword{internal}
