% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSession.R
\name{runSession}
\alias{runSession}
\title{Create a running session.}
\usage{
runSession(pace, heart.rate, profile = NA)
}
\arguments{
\item{pace}{numeric vector giving pace by segment, in seconds}

\item{heart.rate}{numeric vector giving heart rate by segment, in beats per minute}

\item{profile}{runner'sprofile}
}
\value{
Running session.
}
\description{
Create a running session.
}
\examples{
runSession(sample(seq(10, 13, by = .2), 10, replace = T), rnorm(10, mean = 150))
}

