% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterfall.R
\name{runWaterfall}
\alias{runWaterfall}
\title{Run a list of functions as waterfall}
\usage{
runWaterfall(tasks = list(NULL), cb = NULL)
}
\arguments{
\item{tasks}{List of functions (anonymous and named) \strong{required}.}

\item{cb}{Anonymous or named function with signature 
\code{cb(error, data)} \strong{optional}.}
}
\value{
If \code{cb} is \code{NULL} the tasks' return values are returned 
in a named list (on error \code{NULL}). If \code{cb} is a function it is 
called upon completion of all tasks and gets passed an error value 
(default \code{NULL}) as first parameter and a named list of the tasks' 
return values (on error \code{NULL}) as second parameter.
}
\description{
\code{runWaterfall} runs its input tasks sequentially, passing each task's 
return value to the next task, and returns either a named list 
(on error \code{NULL}) or the value of a given callback.
}
\details{
All tasks except the first must have at least one parameter. 
If an error is encountered while calling the tasks without a 
callback \code{runWaterfall} immediately stops execution and returns 
\code{NULL}. If an error is encountered and a callback is defined 
\code{runWaterfall} immediately stops execution and calls the callback with 
the \code{data} parameter set to \code{NULL} and the \code{error} parameter 
set to the encountered error. Thus, the callback will always have only one 
non-\code{NULL} parameter. Within the callback simply check for an error 
with \code{is.null(error)}. If the \code{error} object is not \code{NULL} 
it has a property \code{$task} indicating the function that failed.
}
\note{
\code{\link{bind}} allows binding parameters to a function.
}
\examples{
callback <- function(err, d) {
  if (is.null(err)) d else stop(err, err$task)
}
runWaterfall(list(function() 1L, 
                  function(a) a + 2L, 
                  function(a) a + 3L), 
             callback)

}
\seealso{
\code{\link{runSeries}} \code{\link{runRace}} 
\code{\link{runParallel}} \code{\link{bind}}
 \url{https://github.com/feross/run-waterfall}
}

