% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runZones.R
\name{runZones}
\alias{runZones}
\title{Create effort zones for runner.}
\usage{
runZones(profile, method = "karvonen")
}
\arguments{
\item{profile}{Runner's profile}

\item{method}{Calculation method ("standard", "karvonen")}
}
\value{
Effort zones.
}
\description{
Create effort zones for runner.
}
\examples{
runZones(runProfile(50, 175))
}

