% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-function-calls.R
\name{search_function_calls}
\alias{search_function_calls}
\title{Find calls to given functions in an expression}
\usage{
search_function_calls(expr, functions)
}
\arguments{
\item{expr}{is the expression in which run the search}

\item{functions}{is a string vector in the form of package:::function_name}
}
\value{
A list of the matching calls, or \code{NULL} if there are none.
}
\description{
Walks \code{expr} recursively and collects the calls to any of \code{functions},
matching both the bare name (\code{fun(...)}) and the namespace-qualified forms
(\code{pkg::fun(...)} and \code{pkg:::fun(...)}).
}
