% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{session1}
\alias{session1}
\title{Sample running session 1.}
\format{A data frame with 31 rows and 2 variables:
\describe{
  \item{pace}{pace, in seconds}
  \item{bpm}{heart rate, in beats per minutes}
  ...
}}
\usage{
session1
}
\description{
A dataset containing pace and heart rate from a running session.
}
\keyword{datasets}

