% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stuggi_running}
\alias{stuggi_running}
\title{Example 10k run}
\format{A data frame with 1060 rows and 8 variables:
\describe{
  \item{pointID}{ID for each track point}
  \item{ele}{elevation above sealevel in meters}
  \item{time}{time (in POSIXct-format)}
  \item{runtime}{time since start (in seconds)}
  \item{distance}{distance since the last trackpoint}
  \item{cumulative.distance}{cumulative distance since the beginnning}
  \item{x}{x-coordinate (longitude (East-West-Dimension))}
  \item{y}{y-coordinate (longitude (North-South-Dimension))}
}}
\usage{
stuggi_running
}
\description{
A 10k run I did in my home town Stuttgart, Germany.
}
\keyword{datasets}
