% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_run.R
\name{summariseRun}
\alias{summariseRun}
\title{SummarizeRun}
\usage{
summariseRun(run = stuggi_running)
}
\arguments{
\item{run}{a run data frame returned by \code{import_run}.}
}
\value{
A list with summary statistics.
\describe{
  \item{total_duration}{Total duration as \code{Formal Duration class}}
  \item{total_duration_seconds}{Total duration in seconds}
  \item{total_duration_hours}{Total duration in hours}
  \item{average_speed_kph}{Average speed in kilometers per hour}
  \item{total_distance}{The total distance in kilometers}
}
}
\description{
Get summary statistics for run
}
\examples{
data(stuggi_running)
summariseRun(run = stuggi_running)
}
