% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\alias{time_to_sec}
\alias{sec_to_time}
\title{Switch between time (hh:mm:ss) and seconds}
\usage{
time_to_sec(x)

sec_to_time(x)
}
\arguments{
\item{x}{numeric seconds or character time}
}
\value{
numeric or character vector
}
\description{
Switch between time (hh:mm:ss) and seconds
}
\examples{

time_to_sec("08:30")

sec_to_time(3700)

}
