% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.r
\name{run}
\alias{run}
\title{Run a simulation function}
\usage{
run(data, fun, fixed_parameters, ...)
}
\arguments{
\item{data}{a data.frame (or tibble, in future)}

\item{fun}{a function}

\item{fixed_parameters}{an environment or list}

\item{...}{additional parameters passed to `fun`}
}
\value{
a data.frame including the columns of data and the return from `fun`
}
\description{
Run a simulation function
}
\details{
`fun` must have signature (a1, a2, <...>, aN, fixed_params, ...) where
         1) N is the number of columns in data (the names of these arguments
         don't matter). Note <...> elides intervening arguments and is NOT
         R's ... parameter! 2) the ... parameter is optional 3) The function
         must return a data.frame.
}
\examples{
growth <- function(n, r, K, b) {
  # Ricker-like growth curve in n = log N
  n  + r - exp(n) / K - b
}
data <- expand.grid(
                    b = seq(0.01, 0.5, length.out=10),
                    K = exp(seq(0.1, 5, length.out=10)),
                    r = seq(0.5, 3.5, length.out=10)
                    )
initial_data = list(N0=0.9, T=5, reps=100)

growth_runner <- function(r, K, b, ic) {
  n0 = ic$N0  
  T = ic$T
  reps = ic$reps
  out <- lapply(1:reps, function(i) {
                  for(t in 1:T) {
                    n0 <- growth(n0, r, K, b)
                    }
                  return(n0)
                    })
  data.frame(n_final = do.call(rbind, out))
}


output <- run(data, growth_runner, initial_data)
head(cbind(data, output))
}

