% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingPlot.R
\name{Run2MovingMap}
\alias{Run2MovingMap}
\title{Run2MovingMap}
\usage{
Run2MovingMap(data = stuggi_running, ...)
}
\arguments{
\item{data}{A run data frame returned by \code{import_run}.}

\item{...}{Further Options passed into Leaflet's \code{addPolylines} function.}
}
\value{
A ggplot object.
}
\description{
Visualize Run2MovingMap
}
\section{Note}{

You may customize this plot by either specifying more options within the function or adding other leaflet layers afterwards.
}

\examples{
data(data = stuggi_running)

run2map(data = stuggi_running)
}
