% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package-code.R
\name{run_test_dir}
\alias{run_test_dir}
\title{Run a testthat directory the way test_check would}
\usage{
run_test_dir(package, path, ...)
}
\arguments{
\item{package}{name of the package under test.}

\item{path}{directory containing the testthat tests.}

\item{...}{passed on to \code{\link[testthat:test_dir]{testthat::test_dir()}}.}
}
\value{
The value of \code{\link[testthat:test_dir]{testthat::test_dir()}}.
}
\description{
Simulates \code{testthat::test_check()}, otherwise \code{\link[testthat:test_dir]{testthat::test_dir()}} might
skip some tests. Based on \code{testthat:::test_package_dir}.
}
