% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package-code.R
\name{run_test_env}
\alias{run_test_env}
\title{Create a package environment with access to all package functions}
\usage{
run_test_env(package)
}
\arguments{
\item{package}{name of the package whose namespace to expose.}
}
\value{
An environment containing all bindings of the package namespace.
}
\description{
Based on \code{testthat:::test_pkg_env}.
}
