% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runr.R
\docType{package}
\name{runr}
\alias{runr}
\alias{runr-package}
\title{runr: Running Functions}
\description{
Run a list of functions as series, waterfall, race or parallel.
}
\section{Exports}{

  \code{\link{runSeries}}
  \code{\link{runWaterfall}}
  \code{\link{runRace}}
  \code{\link{runParallel}}
}

\section{Repository}{

  \url{https://github.com/chiefBiiko/runr}
}

